+++ BigBaseV2/src/services/gta_data/gta_data_service.cpp
@@ -5,66 +5,6 @@
 
 namespace big
 {
-	vehicle_item::vehicle_item()
-	{
-		this->name = "";
-		this->display_name = "";
-		this->display_manufacturer = "";
-		this->clazz = "";
-		this->hash = 0;
-	}
-	vehicle_item::vehicle_item(nlohmann::json& item_json)
-	{
-		this->name = item_json["Name"];
-		this->display_name = item_json["Name"];
-		this->display_manufacturer = "";
-		this->clazz = "";
-		this->hash = item_json["Hash"];
-
-		if (!item_json["DisplayName"].is_null())
-		{
-			this->display_name = item_json["DisplayName"];
-		}
-
-		if (!item_json["ManufacturerDisplayName"].is_null())
-		{
-			this->display_manufacturer = item_json["ManufacturerDisplayName"];
-		}
-		else if (!item_json["Manufacturer"].is_null())
-		{
-			this->display_manufacturer = item_json["Manufacturer"];
-		}
-		
-		if (!item_json["Class"].is_null())
-		{
-			this->clazz = item_json["Class"];
-			std::transform(this->clazz.begin(), this->clazz.end(), this->clazz.begin(), ::toupper);
-
-			if (this->clazz == "COMPACTS")
-			{
-				this->clazz = "COMPACT";
-			}
-		}
-	}
-
-	ped_item::ped_item()
-	{
-		this->name = "";
-		this->ped_type = "";
-		this->hash = 0;
-	}
-	ped_item::ped_item(nlohmann::json& item_json)
-	{
-		this->name = item_json["Name"];
-
-		this->ped_type = item_json["Pedtype"];
-		std::transform(this->ped_type.begin(), this->ped_type.end(), this->ped_type.begin(), ::toupper);
-
-		this->hash = item_json["Hash"];
-	}
-
-
-
 	gta_data_service::gta_data_service() :
 		m_vehicle_file(g_file_manager->get_project_file("./lib/vehicles.json")),
 		m_vehicle_file_etag(g_file_manager->get_project_file("./lib/vehicles_etag.txt")),
